refactor: extract admin-scoped storage routes from create_app closure
Move the admin storage view (enriched file listing, cascade delete,
move-to-folder, rename, copy-from) out of the create_app closure into
routes/admin_storage.py, following the existing routes/ extraction
pattern. Drops the now-dead GZIP_UPLOAD_EXTS/DIRECT_UPLOAD_THRESHOLD_BYTES/
content_encoding_for/parse_move_body/parse_rename_body/format_label
closure aliases whose only callers just moved out. Route table is
unchanged in content and order (161 routes before and after); create_app
now only wires lifespan, middleware, auth, background tasks and the
still-inline audit/convert/component/catalog routes.
